Decoupled auth specific functions from component and added initial test coverage
