+++ package.json
@@ -108,6 +108,7 @@
     "gh-pages": "^3.2.3",
     "html-webpack-plugin": "5",
     "identity-obj-proxy": "^3.0.0",
+    "jest-location-mock": "^1.0.9",
     "prettier": "^2.6.2",
     "prettier-eslint": "^10.1.0",
     "prettier-eslint-cli": "^5.0.1",
